@@ -8,19 +8,20 @@ import {
   OnInit,
   Output
   } from '@angular/core';
-import { debounce } from 'rxjs/operator/debounce';
+
 import { DirectoryData } from './directory-data.service';
 import { DirectoryItem } from '../entities/directory-item';
 import { FormControl } from '@angular/forms';
 import { GroupedItems } from '../entities/grouped-items';
 import { I18n } from '../entities/i18n';
 import { Industry } from '../entities/industry';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import {share, filter,take,map,debounceTime} from 'rxjs/operators';
 
 @Component({
   selector: 'app-directory',
   templateUrl: './directory.component.html',
-  styleUrls: ['./directory.component.scss']
+  styleUrls: ['./directory.component.scss'] 
 })
 export class DirectoryComponent {
   groups$: Observable<GroupedItems[]>;
@@ -40,19 +41,19 @@ export class DirectoryComponent {
   ) {
     this.industries$ = data.industries$;
     this.i18n$ = data.i18n$;
-    this.groups$ = data.groupsFilteredByRoute$(this.route).share();
+    this.groups$ = data.groupsFilteredByRoute$(this.route).pipe(share());
 
     this.azList = this.groups$
-      .filter(grp => grp.length > 0) // skip empty sets
-      .take(1) // stop after the first delivery, to not change letters as the filters change
-      .map(grp => this.alphabet.map(letr => Object.assign(new AZLetter(), {
+      .pipe(filter(grp => grp.length > 0)) // skip empty sets
+      .pipe(take(1)) // stop after the first delivery, to not change letters as the filters change
+      .pipe(map(grp => this.alphabet.map(letr => Object.assign(new AZLetter(), {
         Letter: letr,
         Active: grp.find(g => g.label.toLocaleLowerCase() === letr)
-      } )));
+      } ))));
 
     // stream of text input in search
     this.term.valueChanges
-      .debounceTime(400)
+      .pipe(debounceTime(400))
       .subscribe(newTerm => this.router.navigate(['/search', newTerm]));
 
     // keep fields updated from url
